+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For details about compatibility between different releases, see the **Commitment
 - State Description fields for adding context to the (rejected, flagged, suspended) state of Users and OAuth clients.
   - This requires a database schema migration (`ttn-lw-stack is-db migrate`) because of the added columns.
 - Searching for gateways by EUI.
+- Searching for users and OAuth clients by state.
 
 ### Changed